--- Areas/Applicant/JobPostController.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace ApplicationTrackingSystem.Areas.Customer
-{
-    [Area("Customer")]
-    public class JobPostController : Controller
-    {
-        public IActionResult Index()
-        {
-            return View();
-        }
-    }
-}
