--- Repository/IRepository/IRepository.cs
@@ -1,24 +1,0 @@
-﻿using System.Linq.Expressions;
-
-namespace ApplicationTrackingSystem.Repository.IRepository
-{
-    public interface IRepository<T> where T : class
-    {
-        void Add(T entity);
-        void Update(T entity);
-        void Remove(T entity);
-        void Remove(int id);
-        void RemoveRange(IEnumerable<T> values);
-        T Get(int id);
-
-        IEnumerable<T> GetAll(
-            Expression<Func<T, bool>> filter = null,
-            Func<IQueryable<T>, IOrderedQueryable<T>> orderBy = null,
-            string includeProperties = null
-            );
-        T FirstorDefault(
-            Expression<Func<T, bool>> filter = null,
-            string includeProperties = null
-            );
-    }
-}
