--- Repository/Repository.cs
@@ -1,74 +1,0 @@
-﻿using ApplicationTrackingSystem.Data;
-using ApplicationTrackingSystem.Repository.IRepository;
-using Microsoft.EntityFrameworkCore;
-using System.Linq.Expressions;
-
-namespace ApplicationTrackingSystem.Repository
-{
-    public class Repository<T> : IRepository<T> where T : class
-    {
-        private readonly ApplicationDbContext _context;
-        internal DbSet<T> dbset;
-
-        public Repository(ApplicationDbContext context)
-        {
-            _context = context;
-            dbset = _context.Set<T>();
-        }
-        public void Add(T entity)
-        {
-            dbset.Add(entity);
-        }
-        public T FirstorDefault(Expression<Func<T, bool>> filter = null, string includeProperties = null)
-        {
-            IQueryable<T> query = dbset;
-            if (filter != null)
-                query = query.Where(filter);
-            if (includeProperties != null)
-            {
-                foreach (var includeProp in includeProperties.Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries))
-                {
-                    query = query.Include(includeProp);
-                }
-            }
-            return query.FirstOrDefault();
-        }
-        public T Get(int id)
-        {
-            return dbset.Find(id);
-        }
-        public IEnumerable<T> GetAll(Expression<Func<T, bool>> filter = null, Func<IQueryable<T>, IOrderedQueryable<T>> orderBy = null, string includeProperties = null)
-        {
-            IQueryable<T> query = dbset;
-            if (filter != null)
-                query = query.Where(filter);
-            if (includeProperties != null)
-            {
-                foreach (var includeProp in includeProperties.Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries))
-                {
-                    query = query.Include(includeProp);
-                }
-            }
-            if (orderBy != null)
-                return orderBy(query).ToList();
-            return query.ToList();
-        }
-        public void Remove(T entity)
-        {
-            dbset.Remove(entity);
-        }
-        public void Remove(int id)
-        {
-            dbset.Remove(Get(id));
-        }
-        public void RemoveRange(IEnumerable<T> values)
-        {
-            dbset.RemoveRange(values);
-        }
-        public void Update(T entity)
-        {
-            _context.ChangeTracker.Clear();
-            dbset.Update(entity);
-        }
-    }
-}
